stop download, cancel

+++ src/main/java/de/p2tools/mtplayer/controller/data/download/DownloadFactoryDelFilmFile.java
@@ -20,6 +20,7 @@ package de.p2tools.mtplayer.controller.data.download;
 import de.p2tools.mtplayer.controller.config.ProgConfig;
 import de.p2tools.mtplayer.controller.starter.DownloadState;
 import de.p2tools.mtplayer.gui.dialog.DeleteFilmFileDialogController;
+import de.p2tools.mtplayer.gui.dialog.DownloadOnlyStopDialogController;
 import de.p2tools.mtplayer.gui.dialog.DownloadStopDialogController;
 import de.p2tools.mtplayer.gui.tools.MTInfoFile;
 import de.p2tools.mtplayer.gui.tools.MTSubtitle;
@@ -27,12 +28,12 @@ import de.p2tools.p2lib.P2LibConst;
 import de.p2tools.p2lib.alert.PAlert;
 import de.p2tools.p2lib.dialogs.dialog.PDialogExtra;
 import de.p2tools.p2lib.tools.log.PLog;
+import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 
 import java.io.File;
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.List;
 
 public class DownloadFactoryDelFilmFile {
@@ -98,60 +99,91 @@ public class DownloadFactoryDelFilmFile {
         boolean delDownload;
         ObservableList<File> fileList;
 
-        ArrayList<DownloadData> foundList = new ArrayList<>();
+        ObservableList<DownloadData> foundDownloadList = FXCollections.observableArrayList();
         for (DownloadData download : downloads) {
             // DELETE: dann können alle Downloads gelöscht werden, ABBRECHEN: dann nur gestartet
             if (delete ||
                     download.isStateStartedWaiting() || download.isStateStartedRun() || download.isStateError()) {
                 // nur dann läuft er
-                foundList.add(download);
+                foundDownloadList.add(download);
             }
         }
 
-        if (foundList.isEmpty()) {
+        if (foundDownloadList.isEmpty()) {
             // gibt nix zu tun
             return false;
         }
 
+        fileList = getFileList(foundDownloadList);
+        if (fileList.isEmpty()) {
+            // dann wird nur nach dem Löschen des Downloads gefragt
+            delDownload = askForDownload(foundDownloadList, delete);
+        } else {
+            // dann auch nach dem Löschen der Dateien fragen
+            delDownload = askForDownloadAndFile(foundDownloadList, fileList, delete);
+        }
+        return delDownload;
+    }
+
+    private static boolean askForDownload(ObservableList<DownloadData> foundDownloadList, boolean delete) {
+        boolean delDownload;
+        if (ProgConfig.DOWNLOAD_ONLY_STOP.getValue() == DownloadState.DOWNLOAD_ONLY_STOP__DELETE) {
+            // DL löschen, Dateien nicht
+            PLog.sysLog("Stop Download: DL löschen");
+            foundDownloadList.forEach(DownloadData::stopDownload);
+            delDownload = true;
+
+        } else {
+            // dann erstmal fragen
+            PLog.sysLog("Stop Download: Erst mal fragen");
+            DownloadOnlyStopDialogController downloadStopDialogController =
+                    new DownloadOnlyStopDialogController(foundDownloadList, delete);
+
+            if (downloadStopDialogController.getState() == PDialogExtra.STATE.STATE_OK) {
+                // dann soll DL gelöscht werden
+                PLog.sysLog("Stop Download: DL löschen");
+                foundDownloadList.forEach(DownloadData::stopDownload);
+                delDownload = true;
+
+            } else {
+                //dann soll nix gemacht werden
+                PLog.sysLog("Stop Download: Abbruch");
+                delDownload = false;
+            }
+        }
+        return delDownload;
+    }
+
+    private static boolean askForDownloadAndFile(ObservableList<DownloadData> foundDownloadList, ObservableList<File> fileList, boolean delete) {
+        boolean delDownload;
         try {
             switch (ProgConfig.DOWNLOAD_STOP.getValue()) {
                 case DownloadState.DOWNLOAD_STOP__DO_NOT_DELETE:
                     // DL löschen, Dateien nicht
                     PLog.sysLog("Stop Download: DL löschen, Dateien nicht");
-                    foundList.forEach(DownloadData::stopDownload);
-
+                    foundDownloadList.forEach(DownloadData::stopDownload);
                     delDownload = true;
                     break;
 
                 case DownloadState.DOWNLOAD_STOP__DELETE_FILE:
                     // DL und Dateien löschen
                     PLog.sysLog("Stop Download: DL und Dateien löschen");
-                    foundList.forEach(DownloadData::stopDownload);
+                    foundDownloadList.forEach(DownloadData::stopDownload);
                     // und jetzt noch Dateien löschen
-                    fileList = getFileList(foundList);
-                    if (!fileList.isEmpty()) {
-                        try {
-                            Thread.sleep(1000);
-                        } catch (Exception ignore) {
-                            System.out.println("================>");
-                        }
-                        deleteFile(fileList);
-                    }
-
+                    deleteFile(fileList);
                     delDownload = true;
                     break;
 
                 default:
                     // dann erstmal fragen
                     PLog.sysLog("Stop Download: Erst mal fragen");
-                    fileList = getFileList(foundList);
                     DownloadStopDialogController downloadStopDialogController =
-                            new DownloadStopDialogController(fileList, delete);
+                            new DownloadStopDialogController(foundDownloadList, fileList, delete);
 
                     if (downloadStopDialogController.getState() == PDialogExtra.STATE.STATE_1) {
                         // dann soll DL und Datei gelöscht werden
                         PLog.sysLog("Stop Download: DL und Dateien löschen");
-                        foundList.forEach(DownloadData::stopDownload);
+                        foundDownloadList.forEach(DownloadData::stopDownload);
                         // und Dateien löschen
                         deleteFile(fileList);
                         delDownload = true;
@@ -159,7 +191,7 @@ public class DownloadFactoryDelFilmFile {
                     } else if (downloadStopDialogController.getState() == PDialogExtra.STATE.STATE_2) {
                         // dann soll nur der DL gelöscht werden
                         PLog.sysLog("Stop Download: Nur DL löschen");
-                        foundList.forEach(DownloadData::stopDownload);
+                        foundDownloadList.forEach(DownloadData::stopDownload);
                         delDownload = true;
 
                     } else {
@@ -219,24 +251,37 @@ public class DownloadFactoryDelFilmFile {
         return delFileList;
     }
 
-    private static boolean deleteFile(List<File> list) {
-        String delFile = "";
-        boolean ret = true;
-        try {
-            for (File f : list) {
-                delFile = f.getAbsolutePath();
-                PLog.sysLog(new String[]{"Datei löschen: ", f.getAbsolutePath()});
-                if (!f.delete()) {
-                    throw new Exception();
-                }
-            }
-        } catch (Exception ex) {
-            ret = false;
-            PAlert.showErrorAlert("Datei löschen",
-                    "Konnte die Datei nicht löschen!",
-                    "Fehler beim löschen von:" + P2LibConst.LINE_SEPARATORx2 + delFile);
-            PLog.errorLog(989754125, "Fehler beim löschen: " + delFile);
+    private static void deleteFile(List<File> list) {
+        // damit man nicht warten muss, Win braucht da eine Gedenkminute :(
+        if (list.isEmpty()) {
+            return;
         }
-        return ret;
+
+        new Thread(() -> {
+            try {
+                Thread.sleep(1000);
+            } catch (Exception ignore) {
+                System.out.println("================>");
+            }
+
+            String delFile = "";
+            try {
+                for (File f : list) {
+                    delFile = f.getAbsolutePath();
+                    PLog.sysLog(new String[]{"Datei löschen: ", f.getAbsolutePath()});
+                    if (!f.delete()) {
+                        throw new Exception();
+                    }
+                }
+            } catch (Exception ex) {
+                final String df = delFile;
+                Platform.runLater(() -> {
+                    PAlert.showErrorAlert("Datei löschen",
+                            "Konnte die Datei nicht löschen!",
+                            "Fehler beim Löschen von:" + P2LibConst.LINE_SEPARATORx2 + df);
+                    PLog.errorLog(989754125, "Fehler beim löschen: " + df);
+                });
+            }
+        }).start();
     }
 }
